SearchEngine/requirements.txt
blinker==1.6.3
build==1.0.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
elastic-transport==8.15.1
elasticsearch==8.16.0
filelock==3.16.1
Flask==3.0.0
fsspec==2024.10.0
huggingface-hub==0.26.3
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
packaging==23.2
pillow==11.0.0
pip-tools==7.3.0
pyproject_hooks==1.0.0
python-dotenv==1.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==2.7.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.1
